Move some classes/enums to different files.  Documentation
